+++ src/components/Profile/Message.tsx
@@ -1,12 +1,11 @@
 import { Button } from '@components/UI/Button';
 import { Profile } from '@generated/types';
 import { Client } from '@xmtp/xmtp-js';
-import { FC, useEffect } from 'react';
+import { FC } from 'react';
 import toast from 'react-hot-toast';
 import { SIGN_WALLET } from 'src/constants';
 import { useAppStore } from 'src/store/app';
-import { useXmtpStore } from 'src/store/xmtp';
-import { useAccount, useSigner } from 'wagmi';
+import { useSigner } from 'wagmi';
 
 interface Props {
   profile: Profile;
@@ -15,9 +14,6 @@ interface Props {
 const Message: FC<Props> = ({ profile }) => {
   const currentProfile = useAppStore((state) => state.currentProfile);
   const { data: signer, isError, isLoading } = useSigner();
-  const { address } = useAccount();
-  const xmtpState = useXmtpStore((state) => state);
-  const { client, setClient, conversations, setConversations, messages, setMessages, setLoading } = xmtpState;
 
   const sendGm = async () => {
     if (!currentProfile || isError || !signer) {
@@ -27,60 +23,13 @@ const Message: FC<Props> = ({ profile }) => {
 
     // TODO(elise): Swap this out with opening the inbox. This is just a prototype.
     try {
-      const message = await client?.sendMessage(profile.ownedBy, 'gm');
-      toast.success(`Messaged successfully! ${message?.content}`);
+      const xmtp = await Client.create(signer);
+      const message = await xmtp.sendMessage(profile.ownedBy, 'gm');
+      toast.success(`Messaged successfully! ${message.content}`);
     } catch (error) {
       toast.error('Error sending message');
     }
   };
-
-  useEffect(() => {
-    const initXmtpClient = async () => {
-      if (signer) {
-        const xmtp = await Client.create(signer);
-        setClient(xmtp);
-      }
-    };
-    initXmtpClient();
-  }, [signer]);
-
-  useEffect(() => {
-    if (!client) {
-      return;
-    }
-
-    async function listConversations() {
-      setLoading(true);
-      const convos = (await client?.conversations?.list()) || [];
-      Promise.all(
-        convos.map(async (convo) => {
-          if (convo.peerAddress !== address) {
-            const newMessages = await convo.messages();
-            messages.set(convo.peerAddress, newMessages);
-            setMessages(new Map(messages));
-            conversations.set(convo.peerAddress, convo);
-            setConversations(new Map(conversations));
-          }
-        })
-      ).then(() => {
-        setLoading(false);
-      });
-    }
-    const streamConversations = async () => {
-      const stream = (await client?.conversations?.stream()) || [];
-      for await (const convo of stream) {
-        if (convo.peerAddress !== address) {
-          const newMessages = await convo.messages();
-          messages.set(convo.peerAddress, newMessages);
-          setMessages(new Map(messages));
-          conversations.set(convo.peerAddress, convo);
-          setConversations(new Map(conversations));
-        }
-      }
-    };
-    listConversations();
-    streamConversations();
-  }, [client]);
 
   return (
     <Button
