+++ src/pages/conversations.tsx
@@ -1,0 +1,3 @@
+import Conversations from '@components/Conversations';
+
+export default Conversations;
